% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{importData}
\alias{importData}
\title{Proteome Discoverer Import}
\usage{
importData(excelFile)
}
\arguments{
\item{excelFile}{The name of the file output by ProteomeDiscoverer}
}
\description{
A function to import the Excel ProteomeDiscoverer file and extract the protein intensities
}
\details{
A third documentation chuck that appears in the details section of the help file
}
\keyword{import}
